refactor(memory-core): split short-term promotion
